Move check on other scene objects outside of objects list component
